requirements.txt
# Core
django==2.2.5
gunicorn==19.9.0
djangorestframework==3.9.1
django_extensions
django-rest-auth==0.9.3
django-allauth==0.39.1


#mysql
mysqlclient==1.4.4
django-mysql